Move TaskTokenEnum to app/Enums/Auth/Token
